+++ src/lib.rs
@@ -13,7 +13,7 @@ mod log;
 
 use std::ffi::CString;
 
-pub use auxiliary::{AuxField, EnumField, FieldType};
+pub use auxiliary::{AuxField, AuxFieldSetExt, EnumField, FieldType};
 pub use compression::{RecordCompression, SignalCompression};
 pub use error::Slow5Error;
 pub use header::{AuxNamesIter, Header, HeaderExt};